Split logic betwen controller and services
